+++ visualization/app/codeCharta/ui/fileExtensionBar/distributionMetricChooser/distributionMetricChooser.component.ts
@@ -1,4 +1,4 @@
-import { Component, ViewEncapsulation } from "@angular/core"
+import { Component } from "@angular/core"
 import { Store } from "@ngrx/store"
 import { CcState } from "../../../codeCharta.model"
 import { setDistributionMetric } from "../../../state/store/dynamicSettings/distributionMetric/distributionMetric.actions"
@@ -7,7 +7,7 @@ import { distributionMetricSelector } from "../../../state/store/dynamicSettings
 @Component({
     selector: "cc-distribution-metric-chooser",
     templateUrl: "./distributionMetricChooser.component.html",
-    encapsulation: ViewEncapsulation.None
+    styleUrls: ["./distributionMetricChooser.component.scss"]
 })
 export class DistributionMetricChooserComponent {
     distributionMetric$ = this.store.select(distributionMetricSelector)