+++ visualization/app/codeCharta/ui/ribbonBar/areaMetricChooser/areaMetricChooser.component.ts
@@ -1,4 +1,4 @@
-import { Component, ViewEncapsulation } from "@angular/core"
+import { Component } from "@angular/core"
 import { Store } from "@ngrx/store"
 import { CcState } from "../../../codeCharta.model"
 import { setAreaMetric } from "../../../state/store/dynamicSettings/areaMetric/areaMetric.actions"
@@ -6,8 +6,7 @@ import { areaMetricSelector } from "../../../state/store/dynamicSettings/areaMet
 
 @Component({
     selector: "cc-area-metric-chooser",
-    templateUrl: "./areaMetricChooser.component.html",
-    encapsulation: ViewEncapsulation.None
+    templateUrl: "./areaMetricChooser.component.html"
 })
 export class AreaMetricChooserComponent {
     areaMetric$ = this.store.select(areaMetricSelector)