+++ visualization/app/codeCharta/ui/ribbonBar/colorMetricChooser/colorMetricChooser.component.ts
@@ -1,4 +1,4 @@
-import { Component, ViewEncapsulation } from "@angular/core"
+import { Component } from "@angular/core"
 import { setColorMetric } from "../../../state/store/dynamicSettings/colorMetric/colorMetric.actions"
 import { colorMetricSelector } from "../../../state/store/dynamicSettings/colorMetric/colorMetric.selector"
 import { isColorMetricLinkedToHeightMetricSelector } from "../../../state/store/appSettings/isHeightAndColorMetricLinked/isColorMetricLinkedToHeightMetric.selector"
@@ -7,14 +7,11 @@ import { CcState } from "../../../codeCharta.model"
 
 @Component({
     selector: "cc-color-metric-chooser",
-    templateUrl: "./colorMetricChooser.component.html",
-    encapsulation: ViewEncapsulation.None
+    templateUrl: "./colorMetricChooser.component.html"
 })
 export class ColorMetricChooserComponent {
     colorMetric$ = this.store.select(colorMetricSelector)
     isColorMetricLinkedToHeightMetric$ = this.store.select(isColorMetricLinkedToHeightMetricSelector)
-    nonDisabledColor = "rgba(0, 0, 0, 0.38)"
-    disabledColor = "rgba(68,68,68, 1)"
 
     constructor(private store: Store<CcState>) {}