+++ visualization/app/codeCharta/ui/ribbonBar/ribbonBarPanel/ribbonBarPanelSettings.component.ts
@@ -1,0 +1,8 @@
+import { Component } from "@angular/core"
+
+@Component({
+    selector: "cc-ribbon-bar-panel-settings",
+    template: "<ng-content></ng-content>",
+    styleUrl: "./ribbonBarPanelSettings.component.scss"
+})
+export class RibbonBarPanelSettingsComponent {}