+++ visualization/app/codeCharta/ui/viewCube/centerMapButton/centerMapButton.component.ts
@@ -1,11 +1,10 @@
-import { Component, ViewEncapsulation } from "@angular/core"
+import { Component } from "@angular/core"
 import { ThreeMapControlsService } from "../../codeMap/threeViewer/threeMapControls.service"
 
 @Component({
     selector: "cc-center-map-button",
     templateUrl: "./centerMapButton.component.html",
-    styleUrls: ["./centerMapButton.component.scss"],
-    encapsulation: ViewEncapsulation.None
+    styleUrls: ["./centerMapButton.component.scss"]
 })
 export class CenterMapButtonComponent {
     constructor(private threeMapControlsService: ThreeMapControlsService) {}